+++ rsass/tests/spec/core_functions/color/hsl/one_arg/no_alpha.rs
@@ -18,7 +18,7 @@ mod clamped {
             assert_eq!(
                 runner().ok("a {b: hsl(0 100% 500%)}\n"),
                 "a {\
-         \n  b: hsl(0deg, 100%, 100%);\
+         \n  b: hsl(0, 100%, 100%);\
          \n}\n"
             );
         }
@@ -27,7 +27,7 @@ mod clamped {
             assert_eq!(
                 runner().ok("a {b: hsl(0 100% -100%)}\n"),
                 "a {\
-         \n  b: hsl(0deg, 100%, 0%);\
+         \n  b: hsl(0, 100%, 0%);\
          \n}\n"
             );
         }
@@ -41,7 +41,7 @@ mod clamped {
             assert_eq!(
                 runner().ok("a {b: hsl(0 500% 50%)}\n"),
                 "a {\
-         \n  b: hsl(0deg, 100%, 50%);\
+         \n  b: hsl(0, 100%, 50%);\
          \n}\n"
             );
         }
@@ -50,7 +50,7 @@ mod clamped {
             assert_eq!(
                 runner().ok("a {b: hsl(0 -100% 50%)}\n"),
                 "a {\
-         \n  b: hsl(0deg, 0%, 50%);\
+         \n  b: hsl(0, 0%, 50%);\
          \n}\n"
             );
         }
@@ -65,7 +65,7 @@ mod in_gamut {
         assert_eq!(
             runner().ok("a {b: hsl(240 100% 50%)}\n"),
             "a {\
-         \n  b: hsl(240deg, 100%, 50%);\
+         \n  b: hsl(240, 100%, 50%);\
          \n}\n"
         );
     }
@@ -78,7 +78,7 @@ mod in_gamut {
             assert_eq!(
                 runner().ok("a {b: hsl(60 60% 50%)}\n"),
                 "a {\
-         \n  b: hsl(60deg, 60%, 50%);\
+         \n  b: hsl(60, 60%, 50%);\
          \n}\n"
             );
         }
@@ -88,7 +88,7 @@ mod in_gamut {
         assert_eq!(
             runner().ok("a {b: hsl(0 100% 50%)}\n"),
             "a {\
-         \n  b: hsl(0deg, 100%, 50%);\
+         \n  b: hsl(0, 100%, 50%);\
          \n}\n"
         );
     }
@@ -98,7 +98,7 @@ fn named() {
     assert_eq!(
         runner().ok("a {b: hsl($channels: 0 100% 50%)}\n"),
         "a {\
-         \n  b: hsl(0deg, 100%, 50%);\
+         \n  b: hsl(0, 100%, 50%);\
          \n}\n"
     );
 }
@@ -115,7 +115,7 @@ mod units {
             assert_eq!(
                 runner().ok("a {b: hsl(0deg 100% 50%)}\n"),
                 "a {\
-         \n  b: hsl(0deg, 100%, 50%);\
+         \n  b: hsl(0, 100%, 50%);\
          \n}\n"
             );
         }
@@ -129,7 +129,7 @@ mod units {
             assert_eq!(
                 runner().ok("a {b: hsl(0 100% 50)}\n"),
                 "a {\
-         \n  b: hsl(0deg, 100%, 50%);\
+         \n  b: hsl(0, 100%, 50%);\
          \n}\n"
             );
         }
@@ -143,7 +143,7 @@ mod units {
             assert_eq!(
                 runner().ok("a {b: hsl(0 50 50%)}\n"),
                 "a {\
-         \n  b: hsl(0deg, 50%, 50%);\
+         \n  b: hsl(0, 50%, 50%);\
          \n}\n"
             );
         }