+++ rsass/tests/spec/core_functions/color/hsl/three_args/w3c/black_to_white_through.rs
@@ -22,17 +22,17 @@ fn blue() {
              \n  step-11: hsl(240, 100%, 100%);\
              \n}\n"),
         "a {\
-         \n  step-1: hsl(240deg, 100%, 0%);\
-         \n  step-2: hsl(240deg, 100%, 10%);\
-         \n  step-3: hsl(240deg, 100%, 20%);\
-         \n  step-4: hsl(240deg, 100%, 30%);\
-         \n  step-5: hsl(240deg, 100%, 40%);\
-         \n  step-6: hsl(240deg, 100%, 50%);\
-         \n  step-7: hsl(240deg, 100%, 60%);\
-         \n  step-8: hsl(240deg, 100%, 70%);\
-         \n  step-9: hsl(240deg, 100%, 80%);\
-         \n  step-10: hsl(240deg, 100%, 90%);\
-         \n  step-11: hsl(240deg, 100%, 100%);\
+         \n  step-1: hsl(240, 100%, 0%);\
+         \n  step-2: hsl(240, 100%, 10%);\
+         \n  step-3: hsl(240, 100%, 20%);\
+         \n  step-4: hsl(240, 100%, 30%);\
+         \n  step-5: hsl(240, 100%, 40%);\
+         \n  step-6: hsl(240, 100%, 50%);\
+         \n  step-7: hsl(240, 100%, 60%);\
+         \n  step-8: hsl(240, 100%, 70%);\
+         \n  step-9: hsl(240, 100%, 80%);\
+         \n  step-10: hsl(240, 100%, 90%);\
+         \n  step-11: hsl(240, 100%, 100%);\
          \n}\n"
     );
 }
@@ -53,17 +53,17 @@ fn cyan() {
              \n  step-11: hsl(180, 100%, 100%);\
              \n}\n"),
         "a {\
-         \n  step-1: hsl(180deg, 100%, 0%);\
-         \n  step-2: hsl(180deg, 100%, 10%);\
-         \n  step-3: hsl(180deg, 100%, 20%);\
-         \n  step-4: hsl(180deg, 100%, 30%);\
-         \n  step-5: hsl(180deg, 100%, 40%);\
-         \n  step-6: hsl(180deg, 100%, 50%);\
-         \n  step-7: hsl(180deg, 100%, 60%);\
-         \n  step-8: hsl(180deg, 100%, 70%);\
-         \n  step-9: hsl(180deg, 100%, 80%);\
-         \n  step-10: hsl(180deg, 100%, 90%);\
-         \n  step-11: hsl(180deg, 100%, 100%);\
+         \n  step-1: hsl(180, 100%, 0%);\
+         \n  step-2: hsl(180, 100%, 10%);\
+         \n  step-3: hsl(180, 100%, 20%);\
+         \n  step-4: hsl(180, 100%, 30%);\
+         \n  step-5: hsl(180, 100%, 40%);\
+         \n  step-6: hsl(180, 100%, 50%);\
+         \n  step-7: hsl(180, 100%, 60%);\
+         \n  step-8: hsl(180, 100%, 70%);\
+         \n  step-9: hsl(180, 100%, 80%);\
+         \n  step-10: hsl(180, 100%, 90%);\
+         \n  step-11: hsl(180, 100%, 100%);\
          \n}\n"
     );
 }
@@ -84,17 +84,17 @@ fn green() {
              \n  step-11: hsl(120, 100%, 100%);\
              \n}\n"),
         "a {\
-         \n  step-1: hsl(120deg, 100%, 0%);\
-         \n  step-2: hsl(120deg, 100%, 10%);\
-         \n  step-3: hsl(120deg, 100%, 20%);\
-         \n  step-4: hsl(120deg, 100%, 30%);\
-         \n  step-5: hsl(120deg, 100%, 40%);\
-         \n  step-6: hsl(120deg, 100%, 50%);\
-         \n  step-7: hsl(120deg, 100%, 60%);\
-         \n  step-8: hsl(120deg, 100%, 70%);\
-         \n  step-9: hsl(120deg, 100%, 80%);\
-         \n  step-10: hsl(120deg, 100%, 90%);\
-         \n  step-11: hsl(120deg, 100%, 100%);\
+         \n  step-1: hsl(120, 100%, 0%);\
+         \n  step-2: hsl(120, 100%, 10%);\
+         \n  step-3: hsl(120, 100%, 20%);\
+         \n  step-4: hsl(120, 100%, 30%);\
+         \n  step-5: hsl(120, 100%, 40%);\
+         \n  step-6: hsl(120, 100%, 50%);\
+         \n  step-7: hsl(120, 100%, 60%);\
+         \n  step-8: hsl(120, 100%, 70%);\
+         \n  step-9: hsl(120, 100%, 80%);\
+         \n  step-10: hsl(120, 100%, 90%);\
+         \n  step-11: hsl(120, 100%, 100%);\
          \n}\n"
     );
 }
@@ -115,17 +115,17 @@ fn purple() {
              \n  step-11: hsl(300, 100%, 100%);\
              \n}\n"),
         "a {\
-         \n  step-1: hsl(300deg, 100%, 0%);\
-         \n  step-2: hsl(300deg, 100%, 10%);\
-         \n  step-3: hsl(300deg, 100%, 20%);\
-         \n  step-4: hsl(300deg, 100%, 30%);\
-         \n  step-5: hsl(300deg, 100%, 40%);\
-         \n  step-6: hsl(300deg, 100%, 50%);\
-         \n  step-7: hsl(300deg, 100%, 60%);\
-         \n  step-8: hsl(300deg, 100%, 70%);\
-         \n  step-9: hsl(300deg, 100%, 80%);\
-         \n  step-10: hsl(300deg, 100%, 90%);\
-         \n  step-11: hsl(300deg, 100%, 100%);\
+         \n  step-1: hsl(300, 100%, 0%);\
+         \n  step-2: hsl(300, 100%, 10%);\
+         \n  step-3: hsl(300, 100%, 20%);\
+         \n  step-4: hsl(300, 100%, 30%);\
+         \n  step-5: hsl(300, 100%, 40%);\
+         \n  step-6: hsl(300, 100%, 50%);\
+         \n  step-7: hsl(300, 100%, 60%);\
+         \n  step-8: hsl(300, 100%, 70%);\
+         \n  step-9: hsl(300, 100%, 80%);\
+         \n  step-10: hsl(300, 100%, 90%);\
+         \n  step-11: hsl(300, 100%, 100%);\
          \n}\n"
     );
 }
@@ -146,17 +146,17 @@ fn red() {
              \n  step-11: hsl(0, 100%, 100%);\
              \n}\n"),
         "a {\
-         \n  step-1: hsl(0deg, 100%, 0%);\
-         \n  step-2: hsl(0deg, 100%, 10%);\
-         \n  step-3: hsl(0deg, 100%, 20%);\
-         \n  step-4: hsl(0deg, 100%, 30%);\
-         \n  step-5: hsl(0deg, 100%, 40%);\
-         \n  step-6: hsl(0deg, 100%, 50%);\
-         \n  step-7: hsl(0deg, 100%, 60%);\
-         \n  step-8: hsl(0deg, 100%, 70%);\
-         \n  step-9: hsl(0deg, 100%, 80%);\
-         \n  step-10: hsl(0deg, 100%, 90%);\
-         \n  step-11: hsl(0deg, 100%, 100%);\
+         \n  step-1: hsl(0, 100%, 0%);\
+         \n  step-2: hsl(0, 100%, 10%);\
+         \n  step-3: hsl(0, 100%, 20%);\
+         \n  step-4: hsl(0, 100%, 30%);\
+         \n  step-5: hsl(0, 100%, 40%);\
+         \n  step-6: hsl(0, 100%, 50%);\
+         \n  step-7: hsl(0, 100%, 60%);\
+         \n  step-8: hsl(0, 100%, 70%);\
+         \n  step-9: hsl(0, 100%, 80%);\
+         \n  step-10: hsl(0, 100%, 90%);\
+         \n  step-11: hsl(0, 100%, 100%);\
          \n}\n"
     );
 }
@@ -177,17 +177,17 @@ fn yellow() {
              \n  step-11: hsl(60, 100%, 100%);\
              \n}\n"),
         "a {\
-         \n  step-1: hsl(60deg, 100%, 0%);\
-         \n  step-2: hsl(60deg, 100%, 10%);\
-         \n  step-3: hsl(60deg, 100%, 20%);\
-         \n  step-4: hsl(60deg, 100%, 30%);\
-         \n  step-5: hsl(60deg, 100%, 40%);\
-         \n  step-6: hsl(60deg, 100%, 50%);\
-         \n  step-7: hsl(60deg, 100%, 60%);\
-         \n  step-8: hsl(60deg, 100%, 70%);\
-         \n  step-9: hsl(60deg, 100%, 80%);\
-         \n  step-10: hsl(60deg, 100%, 90%);\
-         \n  step-11: hsl(60deg, 100%, 100%);\
+         \n  step-1: hsl(60, 100%, 0%);\
+         \n  step-2: hsl(60, 100%, 10%);\
+         \n  step-3: hsl(60, 100%, 20%);\
+         \n  step-4: hsl(60, 100%, 30%);\
+         \n  step-5: hsl(60, 100%, 40%);\
+         \n  step-6: hsl(60, 100%, 50%);\
+         \n  step-7: hsl(60, 100%, 60%);\
+         \n  step-8: hsl(60, 100%, 70%);\
+         \n  step-9: hsl(60, 100%, 80%);\
+         \n  step-10: hsl(60, 100%, 90%);\
+         \n  step-11: hsl(60, 100%, 100%);\
          \n}\n"
     );
 }