+++ rsass/tests/spec/libsass_closed_issues/issue_1452.rs
@@ -17,7 +17,8 @@ fn test() {
         "Error: () isn\'t a valid CSS value.\
          \n  ,\
          \n2 |   foo: foo(());\
-         \n  |            ^^\
+         \n  |            ^^ value\
+         \n  |        ======= unknown function treated as plain CSS\
          \n  \'\
          \n  input.scss 2:12  root stylesheet",
     );