+++ rsass/tests/spec/values/mod.rs
@@ -18,3 +18,5 @@ mod lists;
 mod maps;
 
 mod numbers;
+
+mod strings;